# docs/openapi_v4.json
{"openapi":"3.1.0","info":{"title":"ECHO Simulator API","description":"Real-time ECHO ecosystem simulation backend","version":"4.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/config":{"get":{"summary":"Get Config","operationId":"get_config_config_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/run":{"post":{"summary":"Run Simulation","operationId":"run_simulation_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunSimulationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run/tournament":{"post":{"summary":"Run Tournament Endpoint","operationId":"run_tournament_endpoint_run_tournament_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TournamentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run/ab":{"post":{"summary":"Run Ab Endpoint","operationId":"run_ab_endpoint_run_ab_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ABExperimentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/run/robustness":{"post":{"summary":"Run Robustness","operationId":"run_robustness_run_robustness_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RobustnessRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ABExperimentRequest":{"properties":{"simulation_config":{"$ref":"#/components/schemas/SimulationConfig","default":{"duration_days":30,"num_creators":6,"num_users":20}},"echo_group":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Echo Group"},"control_group":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Control Group"},"n_rounds":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"N Rounds","default":10}},"type":"object","title":"ABExperimentRequest"},"CreatorConfig":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type","default":"平衡型"},"asset_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Asset Config"},"initial_work":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Initial Work"}},"type":"object","required":["name"],"title":"CreatorConfig"},"EnvironmentParams":{"properties":{"platform_fee_percent":{"type":"number","maximum":50.0,"minimum":0.0,"title":"Platform Fee Percent","default":5},"discovery_algorithm":{"type":"string","title":"Discovery Algorithm","default":"trending"},"market_growth_rate":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Market Growth Rate","default":0.02},"cite_fee":{"type":"number","title":"Cite Fee","default":1.0},"remix_split_default":{"type":"number","title":"Remix Split Default","default":0.3}},"type":"object","title":"EnvironmentParams"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"RobustnessRequest":{"properties":{"simulation_config":{"$ref":"#/components/schemas/SimulationConfig","default":{"duration_days":30,"num_creators":6,"num_users":20}},"parameter":{"type":"string","title":"Parameter"},"range":{"items":{"type":"number"},"type":"array","title":"Range"},"n_runs_per_value":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"N Runs Per Value","default":10},"baseline_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Baseline Config"}},"type":"object","required":["parameter","range"],"title":"RobustnessRequest"},"RunSimulationRequest":{"properties":{"simulation_config":{"$ref":"#/components/schemas/SimulationConfig","default":{"duration_days":30,"num_creators":6,"num_users":20}},"creators":{"anyOf":[{"items":{"$ref":"#/components/schemas/CreatorConfig"},"type":"array"},{"type":"null"}],"title":"Creators"},"users":{"anyOf":[{"items":{"$ref":"#/components/schemas/UserConfig"},"type":"array"},{"type":"null"}],"title":"Users"},"environment_params":{"anyOf":[{"$ref":"#/components/schemas/EnvironmentParams"},{"type":"null"}],"default":{"platform_fee_percent":5.0,"discovery_algorithm":"trending","market_growth_rate":0.02,"cite_fee":1.0,"remix_split_default":0.3}}},"type":"object","title":"RunSimulationRequest"},"SimulationConfig":{"properties":{"duration_days":{"type":"integer","maximum":365.0,"minimum":1.0,"title":"Duration Days","default":30},"num_creators":{"type":"integer","maximum":50.0,"minimum":1.0,"title":"Num Creators","default":6},"num_users":{"type":"integer","maximum":200.0,"minimum":1.0,"title":"Num Users","default":20},"seed":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Seed"}},"type":"object","title":"SimulationConfig"},"TournamentRequest":{"properties":{"simulation_config":{"$ref":"#/components/schemas/SimulationConfig","default":{"duration_days":30,"num_creators":6,"num_users":20}},"strategies":{"items":{"type":"string"},"type":"array","title":"Strategies","default":["保守型","开放型","平衡型","策略型"]},"metrics":{"items":{"type":"string"},"type":"array","title":"Metrics","default":["revenue","satisfaction","ecosystem_health","asset_growth"]},"n_rounds":{"type":"integer","maximum":100.0,"minimum":1.0,"title":"N Rounds","default":10}},"type":"object","title":"TournamentRequest"},"UserConfig":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role","default":"浏览者"},"preferences":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Preferences"},"personality":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Personality"},"initial_wallet":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Initial Wallet"}},"type":"object","required":["name"],"title":"UserConfig"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}